Número Positivo ou Negativo

+++ EstruturaSeleção/questao_01.py
@@ -1,7 +1,7 @@
 
 # 1. Faça um programa que peça um valor e mostre na tela se o valor é positivo ou negativo.
 
-numero = int(input('Digite um número: '))
+numero = float(input('Digite um número: '))
 
 if numero >= 0:
     print('O número {} é positivo'.format(numero))
